refactor PodSetInfo flow to enable immutability
Store the PodSetInfos provided by Kueue's RunWithPodSetInfo method in
a separate array in the AppWrapperComponent instead of immediately
injecting them into the raw templates.  This simplifies both the
Workload controller (which we hope to upstream to Kueue) and the
enforcement of RBAC by allowing us to require that Component.Template
is immutable.